email verification popup (#116)
* Basic draft for email verification popup

* removed emailVerified from redux and skip btn outlined

* Added check verified status button

* Also added EmailVerify component after register

* Moved auth pages to own subfolder

* fix for safari

* Apply suggestions from code review


* fix: register on verification level low
