+++ src/screens/Auth/EmailVerify/EmailVerify.scss
@@ -1,0 +1,35 @@
+@import "../../../colors";
+
+.email-verify {
+  max-width: 250px;
+
+  h2 {
+    margin-bottom: 10px;
+  }
+
+  p {
+    margin-bottom: 10px;
+  }
+
+  .check-mark-icon {
+    font-size: 75px;
+    color: $color-green;
+  }
+
+  .login-footer button {
+    margin-bottom: 10px;
+  }
+
+  .dont-received {
+    margin-top: 10px;
+
+    span {
+      color: $color-primary;
+      cursor: pointer;
+    }
+
+    span:hover {
+      text-decoration: underline;
+    }
+  }
+}
