+++ src/screens/Auth/Login/Login.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .login-form {
   margin: auto;
